YYYYYEEEEEHHHHHAAAAWWWWW
Ok.  So I think I finally solved the problem with FindFirstFileW.  There
were two problems.  One was being hit and I think one was waiting to be
hit:

1. TCHAR is either 1 byte or 2 bytes depending on use.  Since I was
calling with a wide (W) method I needed 2 bytes but I misread this and
needing 1 byte chars.  So now it is unsigned short (which is what wchar_t
seems to be on windows).
2. This one is more sketchy but I found two mentions that MAX_PATH in this
struct can still overflow even though we are honoring the contract of
the struct and the functions which use this struct.  One is in rare cases
you can try to findfirstfile on a reparsepoint and it can return
>MAX_PATH.   The second implied that since NTFS can accept paths up to
32767 that you might somehow get that pushed back into the struct.

So #2 is pretty sketchy but since we don't actually use these fields I
think it is ok to allocate enough memory to guarantee we never overflow
to be worth the cost.