workaround to get AIX flock working
AIX flock lives in libbsd instead of libc, so it couldn't work the way
it does in other platforms (which usually expect it in libc).  This is a
quick workaround that uses fcntl locks instead.  This would not be
sufficient in Linux, because flock and fcntl locks do not interact, but
they do in AIX, so this wrapper works as a platform-specific shim.

As well as the workaround, this patch also implements the standard
'struct flock' POSIX interface and the AIX specific structure, which
might be generally useful elsewhere, though platform specific structures
will have to be defined for other platforms that might want to use it.