Isolate the crypt function into its own library.
See jruby/jruby#5447

musl-libc based Linux distributions do not ship the crypt library,
and since jnr-ffi is set up to error if any requested libraries
fail to load, this prevents native access from loading properly.
For purposes of jruby/jruby#5447, it appears the loading process
succeeds but the crypt function is not bound correctly and
segfaults. This logic tries the various possible names for the
crypt library in turn, eventually trying libc and letting it hard
error if the crypt function cannot be found and bound.