Fix ancillary data recieving on Linux
Receiving more than one ancillary data message on Linux via the recvmsg system
call was incorrect, as it was not correctly calculating the correct offset
for each of the ancillary data messages.  Correctly calculate where the next
message will be, as well as add a test to ensure that we will receive the
data appropriately.