Implement working path for fcntl int...
This commit is part of work for #144.

The fcntl function is defined in C as receiving varargs, which
some commands require and others do not. Which commands require
an argument and which ones don't is not made clear in most
forms of documentation.

In this case, fcntl with no argument failed on Linux PPC64 but
worked fine on Linux x86 and Darwin x86 whe attempting to F_DUPFD.
Since this behavior is undefined we remove testing for it. This
resolves one issue from #144.

The second issue involves the deprecated variadic-int form of
POSIX.fcntl. Whenever it was added, it was never properly hooked
up to jnr-ffi variadic invocation, resulting in the int... form
actually not passing any additional arguments at all. By adding
a single-int-arg form of fcntl, future recompiles will pick up the
appropriate method. By making the variadic form at least support
a single int argument, we fix existing callers that are doing this
common case and not getting the results they expect. Either of
these outcomes would resolve the other fcntl test failure in #144.

Note that both tests have passed repeatedly on Linux x86 (glibc)
and Darwin x86 (10.14), as well as multiple Linux x86 CI machines,
without exposing this issue.

This commit also adds a hard error to the deprecated fcntl form
for all cases except single-int-arg. Better to have a hard error
on this long-deprecated and highly broken form.

The additional deprecation of fcntl with int... is to emphasize to
future contributors that this form should be avoided.