Update jnr-ffi
This includes a fix to make the "fail immediately" option apply
properly when using reflective proxy generation. Previously, the
crypt function would fail to try other libraries because the
symbol lookup was deferred until after the cascading library
logic. This caused some failures in the WEBrick tests for
jruby/jruby##6171.