Use tricks to improve module visibility issues
When open access to java.io is not granted to jnr-posix, display
a warning about the APIs that will be broken.

When open access is granted but jnr-posix is in the unnamed
module, explicitly open the packages to eliminate the warnings.

See https://github.com/jnr/jnr-posix/wiki/Using-jnr%E2%80%90posix-with-Java-Modules

Eliminates warnings from unnamed module open access requests on
releases prior to Java 17 and provides better information (with
link to the above wiki page) on Java 17+ for unnamed modules or
when open access is denied to the jnr-posix module.

See discussion in https://github.com/jnr/jnr-posix/issues/190